Create project structure
This is adapted from my 2020 Advent of Code, which I also did in Rust. I
wish there were a cleaner/nicer way to structure the entrypoints for
each of the days (maybe if I put them in `bin/` ??). I rewrote my Python
input-fetching logic in Rust. I didn't bother to add support for
submitting my answer from the program itself, but maybe I should
consider that.

I have not touched Rust in quite a while! It doesn't come naturally, but
that's why I'm using it this year.