Day 7 part 2 (as submitted)
This one took me a while. My first approach involved explicitly tracking
each of the paths, which was futile because of how many things I was
trying to store uniquely.

My second approach was this, but I didn't think to increment the
existing value when below was not a split. It took me a while (and a
visualization) to find that bug, but once I fixed it, it worked.

I don't know why I wrote a generic function to do what in Python would
be `d[k] += v`. I also feel that it's likely that that functionality
already exists in the Rust stdlib.