Day 9 part 2 (as submitted)
Wowee! I was just working with big numbers and figuring out the right
(and wrong) ways to iterate over a bunch of things. It's much better to
check sparse points (even 600k of them) for presence in a bounding
rectangle than to check every point of that rectangle for membership in
a set.

This still takes like 10 seconds in release mode…